+++ src/main/kotlin/com/example/Employee/services/EmployeeService.kt
@@ -1,5 +1,7 @@
 package com.example.Employee.services
 
+import com.example.Employee.exceptions.EntityNotFoundException
+import com.example.Employee.exceptions.InvalidRequestException
 import com.example.Employee.models.Employee
 
 import com.example.Employee.repositories.EmployeeRepository
@@ -10,7 +12,12 @@ import java.util.*
 @Service
 public class EmployeeService (private val employeeRepository: EmployeeRepository){
     fun addDetails(employee: Employee){
-        employeeRepository.save(employee)
+        if(employee.age<18){
+            throw InvalidRequestException("Age should be above 18")
+        }
+        else {
+            employeeRepository.save(employee)
+        }
     }
 
     fun get(): List<Employee> {
@@ -18,16 +25,31 @@ public class EmployeeService (private val employeeRepository: EmployeeRepository
     }
 
     fun getEmployee(employeeId: Int): Employee {
-        return employeeRepository.findById(employeeId).get()
+        return employeeRepository.findById(employeeId).orElseThrow{EntityNotFoundException("employee not found")}
     }
 
     fun deleteEmployee(employeeId: Int){
-        employeeRepository.deleteById(employeeId)
+        val employee = employeeRepository.findById(employeeId)
+        if(employee.isPresent) {
+            employeeRepository.deleteById(employeeId)
+        } else{
+            throw EntityNotFoundException("employee not found")
+        }
     }
 
     fun updateEmployee(employeeId: Int, employee: Employee) {
-        employee.employeeId=employeeId
-        employeeRepository.save(employee)
+        val employeeExists = employeeRepository.findById(employeeId)
+        if(employeeExists.isPresent) {
+            employee.employeeId = employeeId
+            if(employee.age>18) {
+                employeeRepository.save(employee)
+            }
+            else{
+                throw InvalidRequestException("Age should be above 18")
+            }
+        } else{
+            throw EntityNotFoundException("employee not found")
+        }
     }
